Refactored Parser.hs
