+++ lib/theory/src/Theory/Text/Parser/Token.hs
@@ -7,6 +7,8 @@
 -- Portability : portable
 --
 -- Tokenizing infrastructure
+{-# LANGUAGE TypeSynonymInstances #-}
+{-# LANGUAGE FlexibleInstances #-}
 module Theory.Text.Parser.Token (
   -- * Symbols
     symbol
@@ -111,6 +113,8 @@ import           Text.Parsec         hiding ((<|>))
 import qualified Text.Parsec.Token   as T
 
 import           Theory
+import qualified Control.Monad.Catch as Catch
+import Data.Functor.Identity
 
 
 ------------------------------------------------------------------------------
@@ -119,6 +123,10 @@ import           Theory
 
 -- | A parser for a stream of tokens.
 type Parser a = Parsec String MaudeSig a
+
+-- We can throw exceptions, but not catch them
+instance Catch.MonadThrow (ParsecT String MaudeSig Data.Functor.Identity.Identity) where
+    throwM e = fail (show e)
 
 -- Use Parsec's support for defining token parsers.
 spthy :: T.TokenParser MaudeSig
@@ -221,7 +229,7 @@ naturalSubscript = T.lexeme spthy $ do
   where
     subscriptDigitToInteger d = toInteger $ fromEnum d - fromEnum '₀'
 
-    
+
 -- | A comma separated list of elements.
 commaSep :: Parser a -> Parser [a]
 commaSep = T.commaSep spthy
