+++ lib/theory/tamarin-prover-theory.cabal
@@ -109,3 +109,14 @@ library
       Theory.Tools.Wellformedness
 
     other-modules:
+      Theory.Text.Parser.Exceptions
+      Theory.Text.Parser.Fact
+      Theory.Text.Parser.Formula
+      Theory.Text.Parser.Lemma
+      Theory.Text.Parser.Let
+      Theory.Text.Parser.Proof
+      Theory.Text.Parser.Restriction
+      Theory.Text.Parser.Rule
+      Theory.Text.Parser.Sapic
+      Theory.Text.Parser.Signature
+      Theory.Text.Parser.Term
